0.0.9 - Added ability to force the inclusion of a library.
This is particularly useful if you have a layout that is including the library at the top of the page while a view is including it at the bottom via a library with a core component. Since the layout is rendered last, the library is never included at the top of the page and that can cause unexpected results.